Modularize notion functions
